kith-platform/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
chromadb==0.4.15
click==8.1.8
coloredlogs==15.0.1
coverage==7.10.2
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
filelock==3.18.0
Flask==2.3.3
flatbuffers==25.2.10
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
# httptools==0.6.4  # Performance dependency for uvicorn - not explicitly used
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
# kubernetes==33.1.0  # Kubernetes client - unused in current implementation
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pluggy==1.6.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==5.29.5
# psycopg2-binary==2.9.7  # PostgreSQL adapter - unused (app uses SQLite)
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
# shellingham==1.5.4  # Shell detection - unused in current implementation
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.21
sympy==1.14.0
tenacity==9.1.2
telethon==1.34.0
tokenizers==0.21.4
tomli==2.2.1
tqdm==4.67.1
# typer==0.16.0  # CLI framework - unused in current implementation
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
# uvloop==0.21.0  # Performance dependency for asyncio - not explicitly used
vobject==0.9.6.1
# watchfiles==1.1.0  # File watching for development - not explicitly used
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
openai==0.28.1
schedule==1.2.2
pyaes==1.6.1
sentry-sdk==2.19.2
PyPDF2==3.0.1
pdfplumber==0.11.7
google-cloud-vision==3.10.2
